Minimal implementation of #683 that maintains API compatibility
